include alternative path fix, parse NaN values, block context menu when focus on, fix texture crop for negative vaules, fix asv file save, loose consist reverse option in context menu, orts weather fix, quad tree faster name id, possible fix for NaN values in TDB, support for more geo terrain data loading

+++ Terrain.cpp
@@ -921,7 +921,7 @@ float Terrain::getErrorBias(){
 }
 
 void Terrain::getWTileIds(QSet<int> &ids){
-    ids.insert(mojex*10000+mojez);
+    ids.insert((int)mojex*(int)10000+(int)mojez);
 }
 
 void Terrain::setErrorBias(float val){
